+++ OrangeDesignSystem/Sources/Components/Slider/ODSSliderView.swift
@@ -23,6 +23,12 @@
 
 import SwiftUI
 
+///
+/// <a href="https://system.design.orange.com/0c1af118d/p/7559da-sliders/b/253eea" target="_blank">ODS Slider</a>.
+///
+/// Based on the native `Slider`, the `ODSSLider` offers to the user the possibility
+/// to type direcly on the slider's track to get a value.
+///
 public struct ODSSlider<V, ValueLabel>: View where V: BinaryFloatingPoint, V.Stride: BinaryFloatingPoint, ValueLabel: View {
     @Binding var value: V
     public let range: ClosedRange<V>
@@ -30,6 +36,17 @@ public struct ODSSlider<V, ValueLabel>: View where V: BinaryFloatingPoint, V.Str
     let minimumValueLabel: () -> ValueLabel
     let maximumValueLabel: () -> ValueLabel
 
+    /// Creates an unlabeled slider to select a value from a given range, subject to a
+    /// step increment.
+    ///
+    /// - Parameters:
+    ///   - value: The selected value within `range`.
+    ///   - range: The range of the valid values.
+    ///   - step: The distance between each valid value. default 1
+    ///
+    /// The `value` of the created instance is equal to the position of
+    /// the given value within `range`.
+    ///
     public init(value: Binding<V>, range: ClosedRange<V>, step: V.Stride = 1) where ValueLabel == EmptyView {
         _value = value
         self.range = range
@@ -42,6 +59,19 @@ public struct ODSSlider<V, ValueLabel>: View where V: BinaryFloatingPoint, V.Str
         }
     }
 
+    /// Creates a slider to select a value from a given range, subject to a
+    /// step increment. which displays the provided labels.
+    ///
+    /// - Parameters:
+    ///   - value: The selected value within `range`.
+    ///   - range: The range of the valid values.
+    ///   - step: The distance between each valid value. default 1
+    ///   - minimumValueLabel: A view that describes `range.lowerBound`.
+    ///   - maximumValueLabel: A view that describes `range.lowerBound`.
+    ///
+    /// The `value` of the created instance is equal to the position of
+    /// the given value within `range`.
+    ///
     public init(value: Binding<V>, range: ClosedRange<V>, step: V.Stride = 1, @ViewBuilder minimumLabelView: @escaping () -> ValueLabel, @ViewBuilder maximumLabelView: @escaping () -> ValueLabel) {
         _value = value
         self.range = range
@@ -53,7 +83,9 @@ public struct ODSSlider<V, ValueLabel>: View where V: BinaryFloatingPoint, V.Str
     public var body: some View {
         VStack {
             HStack(alignment: .center) {
+
                 minimumValueLabel()
+
                 GeometryReader { geometry in
                     Slider(
                         value: $value,
@@ -70,6 +102,7 @@ public struct ODSSlider<V, ValueLabel>: View where V: BinaryFloatingPoint, V.Str
                             height: geometry.size.height,
                             alignment: .center)
                 }
+
                 maximumValueLabel()
             }
         }