+++ OrangeDesignSystem/Sources/Guidelines/Spacing/ODSSpacing.swift
@@ -24,12 +24,22 @@
 import Foundation
 import SwiftUI
 
+///
+/// Spacing are used to position elements on the screen and layout within components.
+///
 public enum ODSSpacing {
+    /// none: 0
     public static let none: Double = 0
+    /// xs: 4
     public static let xs: Double = 4
+    /// s: 8
     public static let s: Double = 8
+    /// m: 16
     public static let m: Double = 16
+    /// l: 24
     public static let l: Double = 24
+    /// xl: 32
     public static let xl: Double = 32
+    /// xxl: 40
     public static let xxl: Double = 40
 }