+++ OrangeDesignSystemDemo/OrangeDesignSystemDemo/Views/Guidelines/Colors/ColorsList.swift
@@ -54,11 +54,11 @@ struct ColorList: View {
                 SectionTitle(title: "Core")
 
                 VStack {
-                    HStack(spacing: 15) {
+                    HStack(spacing: ODSSpacing.m) {
                         ColorBig(color: ODSColor.coreOrange, bordered: false)
                         ColorBig(color: ODSColor.coreWhite, bordered: screenState.colorScheme == .light)
                     }
-                    HStack(spacing: 15) {
+                    HStack(spacing: ODSSpacing.m) {
                         ColorBig(color: ODSColor.coreBlack, bordered: screenState.colorScheme == .dark)
                         ColorBig(color: ODSColor.coreObsGrey, bordered: false)
                     }
@@ -69,11 +69,11 @@ struct ColorList: View {
                     .padding(.top, ODSSpacing.l)
 
                 VStack {
-                    HStack(spacing: 15) {
+                    HStack(spacing: ODSSpacing.m) {
                         ColorBig(color: ODSColor.functionalPositive, bordered: false)
                         ColorBig(color: ODSColor.functionalNegative, bordered: false)
                     }
-                    HStack(spacing: 15) {
+                    HStack(spacing: ODSSpacing.m) {
                         ColorBig(color: ODSColor.functionalAlert, bordered: false)
                         ColorBig(color: ODSColor.functionalInfo, bordered: false)
                     }
@@ -84,33 +84,34 @@ struct ColorList: View {
                     .padding(.top, ODSSpacing.l)
 
                 VStack {
-                    HStack(spacing: 15) {
+                    HStack(spacing: ODSSpacing.m) {
                         ColorSmall(color: ODSColor.supportingBlue100)
                         ColorSmall(color: ODSColor.supportingBlue200)
                         ColorSmall(color: ODSColor.supportingBlue300)
                     }
-                    HStack(spacing: 15) {
+                    HStack(spacing: ODSSpacing.m) {
                         ColorSmall(color: ODSColor.supportingYellow100)
                         ColorSmall(color: ODSColor.supportingYellow200)
                         ColorSmall(color: ODSColor.supportingYellow300)
                     }
-                    HStack(spacing: 15) {
+                    HStack(spacing: ODSSpacing.m) {
                         ColorSmall(color: ODSColor.supportingGreen100)
                         ColorSmall(color: ODSColor.supportingGreen200)
                         ColorSmall(color: ODSColor.supportingGreen300)
                     }
-                    HStack(spacing: 15) {
+                    HStack(spacing: ODSSpacing.m) {
                         ColorSmall(color: ODSColor.supportingPink100)
                         ColorSmall(color: ODSColor.supportingPink200)
                         ColorSmall(color: ODSColor.supportingPink300)
                     }
-                    HStack(spacing: 15) {
+                    HStack(spacing: ODSSpacing.m) {
                         ColorSmall(color: ODSColor.supportingPurple100)
                         ColorSmall(color: ODSColor.supportingPurple200)
                         ColorSmall(color: ODSColor.supportingPurple300)
                     }
                 }
-                .padding(EdgeInsets(top: ODSSpacing.xs, leading: ODSSpacing.m, bottom: ODSSpacing.xs, trailing: ODSSpacing.m))
+                .padding(.vertical, ODSSpacing.xs)
+                .padding(.horizontal, ODSSpacing.m)
             }
             .background(screenState.colorScheme == .light ? Color.white : Color.black)
         }
@@ -131,7 +132,8 @@ struct SectionTitle: View {
     @EnvironmentObject var screenState: ScreenState
 
     var body: some View {
-        Text(title).odsFont(style: .title1)
+        Text(title)
+            .odsFont(.title1)
             .foregroundColor(screenState.colorScheme == .dark ? Color.white : Color.black)
             .frame(maxWidth: .infinity, alignment: .leading)
             .padding(EdgeInsets(top: ODSSpacing.s, leading: ODSSpacing.m, bottom: ODSSpacing.xs, trailing: ODSSpacing.m))