+++ OrangeDesignSystemDemo/Pages/ProgressIndicatorPage.swift
@@ -32,7 +32,7 @@ struct ProgressIndicatorPage: View {
                 .resizable()
                 .aspectRatio(contentMode: .fill)
 
-            VStack(alignment: .leading, spacing: 20) {
+            VStack(alignment: .leading, spacing: ODSSpacing.l) {
                 ComponentDescription(text: "Progress indicators show users that elements or pages are loading.")
                 VariantsTitle()
                 ProgressBar()
@@ -57,7 +57,7 @@ struct ProgressBar: View {
 
     var body: some View {
         Text("Progress bar")
-            .odsFont(style: .title2)
+            .odsFont(.title2)
         ProgressView("Downloading…", value: secondsElapsed, total: maxSeconds)
             .onReceive(timer) { _ in
                 if secondsElapsed < maxSeconds {
@@ -72,7 +72,7 @@ struct ProgressBar: View {
 struct ProgressIndicator: View {
     var body: some View {
         Text("Progress indicator")
-            .odsFont(style: .title2)
+            .odsFont(.title2)
         HStack {
             Spacer()
             ProgressView()